requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.0
requests==2.31.0
google-generativeai==0.5.4
gitpython==3.1.43
pydantic==2.6.4